refactor: Use u32 instead of u8 for faster DP calculations (#128)
* refactor: Use u32 instead of u8 for faster DP calculations

* refactor: Fix clippy::unnecessary_cast

+++ src/standard/common.rs
@@ -13,9 +13,9 @@ pub(super) struct Entry {
 
 #[inline]
 pub(super) fn update_min(
-    lhs_number: &mut u8,
+    lhs_number: &mut u32,
     lhs_tiles: &mut TileFlags,
-    rhs_number: u8,
+    rhs_number: u32,
     rhs_tiles: TileFlags,
 ) {
     match (*lhs_number).cmp(&rhs_number) {
